[GTK] Don't leak GStaticRecMutex https://bugs.webkit.org/show_bug.cgi?id=104901
Patch by Alberto Garcia <agarcia@igalia.com> on 2012-12-13
Reviewed by Martin Robinson.

The mutex in WebKitWebAudioSourceGStreamer is allocated using
g_new(), but that memory is not freed when the object is
destroyed.

This patch replaces that pointer with the structure itself, which
is faster and doesn't need to be explicitly freed.

* platform/audio/gstreamer/WebKitWebAudioSourceGStreamer.cpp:
(_WebKitWebAudioSourcePrivate):
(webkit_web_audio_src_init):
(webKitWebAudioSrcFinalize):

Identifier: 123132@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@137587 268f45cc-cd09-0410-ab3c-d52691b4dbfc