[GTK] accessibiltiy/aria-hidden.html is failing https://bugs.webkit.org/show_bug.cgi?id=98352
Patch by Joanmarie Diggs <jdiggs@igalia.com> on 2012-12-13
Reviewed by Martin Robinson.

The test is failing due to platform differences in the implementation of
AccessibilityUIElement::allAttributes(). This test does not require this
method as it is only looking at the AXTitle attribute.

* accessibility/aria-hidden-expected.txt: Included the actual reason for
passing/failing in the results output.
* accessibility/aria-hidden.html: Included the actual reason for passing/
failing in the results output.
* platform/gtk/TestExpectations: Unskipped the failing test.

Identifier: 123159@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@137619 268f45cc-cd09-0410-ab3c-d52691b4dbfc