https://bugs.webkit.org/show_bug.cgi?id=102599 ASSERT loading Acid3 test in run-safari --debug (r135050)
Reviewed by Simon Fraser.

New flag IncludeCompositedDescendants will always calculate the layer bounds for
descendants, even when the are composited.
* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::calculateLayerBounds):
* rendering/RenderLayer.h:

calculateCompositedBounds() should not include this new flag, so instead of
calling that, call calculateLayerBounds() directly with the
IncludeCompositedDescendants flag. This will get us the information we need
upfront and avoid the toggling later on.
* rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::requiresCompositingForPosition):



Identifier: 123213@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@137690 268f45cc-cd09-0410-ab3c-d52691b4dbfc