ChildNodesLazySnapshot::nextNode() can crash in Debug builds https://bugs.webkit.org/show_bug.cgi?id=104982
Reviewed by Ryosuke Niwa.

The original fix in r137702 was not enough. r137702 fixed a crash
in Release builds but didn't fix a crash in Debug builds.
We have to also change a return type of ChildNodesLazySnapshot::nextNode()
from a raw pointer to a RefPtr.

Test: fast/dom/insertedIntoDocument-no-crash.html
(Confirm that this test does not crash in Debug builds.)

* dom/ContainerNode.h:
(WebCore::ChildNodesLazySnapshot::nextNode):
* dom/ContainerNodeAlgorithms.cpp:
(WebCore::ChildNodeInsertionNotifier::notifyDescendantInsertedIntoDocument):
(WebCore::ChildNodeRemovalNotifier::notifyDescendantRemovedFromDocument):


Identifier: 123260@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@137739 268f45cc-cd09-0410-ab3c-d52691b4dbfc