Remove the unneeded NetworkProcessCrashed message https://bugs.webkit.org/show_bug.cgi?id=105114
Reviewed by Sam Weinig.

Remove the NetworkProcessCrashed message; it's not needed since the connection member variable is already
nulled out in WebProcess::networkProcessConnectionClosed which is called when the Connection::Client::didClose member
function is called on the connection between the web process and network process.

Furthermore, this message was handled on the connection queue which runs on a different thread and the handler was not
thread-safe so this could in theory cause bad crashes.

* UIProcess/Network/NetworkProcessProxy.cpp:
(WebKit::NetworkProcessProxy::didClose):
* WebProcess/WebProcess.cpp:
* WebProcess/WebProcess.h:
(WebProcess):
* WebProcess/WebProcess.messages.in:

Identifier: 123328@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@137815 268f45cc-cd09-0410-ab3c-d52691b4dbfc