[EFL] Allow the build system to detect EGL support. https://bugs.webkit.org/show_bug.cgi?id=105287
Patch by Kondapally Kalyan <kalyan.kondapally@intel.com> on 2012-12-18
Reviewed by Laszlo Gombos.

Currently, we don't have any way to determine if EGL is supported by the build.
This patch adds support for this. The patch doesn't make any changes to take this into
use. This will be done in another patch.

* Source/cmake/FindEGL.cmake:

Identifier: 123537@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@138033 268f45cc-cd09-0410-ab3c-d52691b4dbfc