https://bugs.webkit.org/show_bug.cgi?id=102579 [mac] Enable scaled cursors
Reviewed by Dean Jackson.

Source/JavaScriptCore:

* Configurations/FeatureDefines.xcconfig:

Source/WebCore:

Not much is required to get this turned on. We just need the feature to be
defined, and we need to initialize m_imageScaleFactor in the Cursor constructors.

* Configurations/FeatureDefines.xcconfig:
* platform/mac/CursorMac.mm:
(WebCore::Cursor::Cursor):
(WebCore::Cursor::operator=):

Source/WebKit/mac:

* Configurations/FeatureDefines.xcconfig:

Source/WebKit2:

* Configurations/FeatureDefines.xcconfig:

LayoutTests:

Un-skip these tests now that the feature is enabled.
* platform/mac/TestExpectations:



Identifier: 123573@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@138069 268f45cc-cd09-0410-ab3c-d52691b4dbfc