Web Inspector: deny access from injected script to nodes from document with another origin https://bugs.webkit.org/show_bug.cgi?id=105423
Reviewed by Pavel Feldman.

Source/WebCore:

Check that calling context can access the document inspected node belong to
before returning JS wrapper for the node.

Test: http/tests/inspector-protocol/access-inspected-object.html

* bindings/js/JSInjectedScriptHostCustom.cpp:
(WebCore::JSInjectedScriptHost::inspectedObject):
* bindings/v8/custom/V8InjectedScriptHostCustom.cpp:
(WebCore::InjectedScriptHost::nodeAsScriptValue):

LayoutTests:

Test that $0 is ineaccible in the main frame if it is a node from an iframe that
is not accassible beacause of cross origin access checks.

* http/tests/inspector-protocol/access-inspected-object-expected.txt: Added.
* http/tests/inspector-protocol/access-inspected-object.html: Added.
* http/tests/inspector-protocol/resources/test-page.html: Added.

Identifier: 123718@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@138228 268f45cc-cd09-0410-ab3c-d52691b4dbfc